move ./functions, homebrew installer, & node installer into ./scripts
