+++ git/template.symlink.gitconfig
@@ -1,5 +1,5 @@
 [user]
-        name = AUTHORNAME
-        email = AUTHOREMAIL
+		name = AUTHORNAME
+		email = AUTHOREMAIL
 [credential]
-        helper = GIT_CREDENTIAL_HELPER
+		helper = GIT_CREDENTIAL_HELPER
