+++ scripts/bootstrap.ts
@@ -9,7 +9,7 @@ import {
 	logSuccess,
 	readFile,
 	searchDirectory
-} from '../functions/index';
+} from './functions/index';
 
 /** ask for git author information to fill out the gitconfig file */
 async function setupGitconfig(): Promise<void> {
@@ -28,10 +28,10 @@ async function setupGitconfig(): Promise<void> {
 		const authorEmail = await input(' - What is your github author email?');
 
 		await exec(
-			`sed`
-			+ `-e "s/AUTHORNAME/${authorName}/g"`
-			+ `-e "s/AUTHOREMAIL/${authorEmail}/g"`
-			+ `-e "s/GIT_CREDENTIAL_HELPER/${gitCredential}/g"`
+			`sed `
+			+ `-e "s/AUTHORNAME/${authorName}/g" `
+			+ `-e "s/AUTHOREMAIL/${authorEmail}/g" `
+			+ `-e "s/GIT_CREDENTIAL_HELPER/${gitCredential}/g" `
 			+ `${gitconfigTemplatePath} > ${gitconfigPath}`
 		);
