+++ scripts/functions/array.ts
