+++ scripts/functions/index.ts
