+++ scripts/functions/link-file.ts
