+++ scripts/functions/logging.sh
