+++ scripts/functions/logging.ts
