+++ scripts/functions/object.ts
