+++ scripts/functions/types.ts
