+++ scripts/functions/variable-access.sh
